Rename node kind file
